Move visualizer to use dagre layout
